+++ packages/web/src/pages/collections/[id].tsx
@@ -150,26 +150,26 @@ export default function CollectionPage() {
                   >
                     {/* Media Content */}
                     <div className="relative aspect-square bg-gray-800">
-                      {post.s3VideoUrl ? (
+                      {post.videoUrl ? (
                         <video
                           className="w-full h-full object-cover"
                           poster={
-                            `${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.s3ThumbnailSrc}` ||
-                            `${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.s3DisplayUrl}` ||
+                            `${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.thumbnailSrc}` ||
+                            `${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.displayUrl}` ||
                             undefined
                           }
                           controls
                           preload="metadata"
                         >
                           <source
-                            src={`${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.s3VideoUrl}`}
+                            src={`${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.videoUrl}`}
                             type="video/mp4"
                           />
                           Your browser does not support the video tag.
                         </video>
-                      ) : post.s3DisplayUrl ? (
+                      ) : post.displayUrl ? (
                         <img
-                          src={`${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.s3DisplayUrl}`}
+                          src={`${process.env.NEXT_PUBLIC_AWS_BUCKET_PROXY_URL}/${post.displayUrl}`}
                           alt="Instagram post"
                           className="w-full h-full object-cover"
                           loading="lazy"
@@ -187,7 +187,7 @@ export default function CollectionPage() {
                       )}
 
                       {/* Video indicator */}
-                      {post.s3VideoUrl && (
+                      {post.videoUrl && (
                         <div className="absolute top-3 right-3 bg-black/60 rounded-full p-1">
                           <svg className="w-4 h-4 text-white" fill="currentColor" viewBox="0 0 20 20">
                             <path d="M2 6a2 2 0 012-2h6l2 2h6a2 2 0 012 2v6a2 2 0 01-2 2H4a2 2 0 01-2-2V6zM14.553 7.106A1 1 0 0014 8v4a1 1 0 00.553.894l2 1A1 1 0 0018 13V7a1 1 0 00-1.447-.894l-2 1z" />
