+++ packages/web/src/server/db/schema.ts
@@ -51,11 +51,8 @@ export const posts = createTable(
     shortcode: d.varchar({ length: 255 }),
     description: d.text(),
     videoUrl: d.text(),
-    s3VideoUrl: d.text(),
     thumbnailSrc: d.text(),
-    s3ThumbnailSrc: d.text(),
     displayUrl: d.text(),
-    s3DisplayUrl: d.text(),
     parsedAt: d.timestamp({ withTimezone: true }),
     createdAt: d.timestamp({ withTimezone: true }).default(sql`CURRENT_TIMESTAMP`).notNull(),
     updatedAt: d.timestamp({ withTimezone: true }).$onUpdate(() => new Date()),
