+++ packages/web/src/server/di/container.ts
@@ -2,8 +2,8 @@ import 'reflect-metadata';
 import { container } from 'tsyringe';
 
 // Import classes to ensure decorators are processed
-import '~/server/features/database/database-service';
-import '~/server/features/s3/s3-service';
+import '~/server/db/database-service';
+import '~/server/s3/s3-service';
 import '~/server/features/collections/collections-service';
 import '~/server/features/jobs/jobs-service';
 import '~/server/features/posts/posts-service';
@@ -14,7 +14,7 @@ import '~/server/queue/setup';
 import '~/server/queue';
 import '~/server/queue/jobs';
 import '~/server/queue/workers';
-import '~/server/queue/handlers/parse-instagram-post';
-import '~/server/queue/handlers/upload-instagram-post';
+import '~/server/features/posts/handlers/parse-instagram-post-handler';
+import '~/server/features/posts/handlers/upload-instagram-post-handler';
 
 export { container };
